ft_substr: changed argument type to 'size_t'

+++ 42-cursus/common-core-rank01/get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: rtorrent <rtorrent@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 11:25:11 by rtorrent          #+#    #+#             */
-/*   Updated: 2023/06/30 10:30:32 by rtorrent         ###   ########.fr       */
+/*   Updated: 2023/07/04 17:52:37 by rtorrent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,6 +49,6 @@ void	ft_lstclear(t_list **plst, void (*del)(void *));
 t_list	*ft_lstnew(void *content);
 char	*ft_strchr(const char *s, int c);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
-char	*ft_substr(const char *s, unsigned int start, size_t len);
+char	*ft_substr(const char *s, size_t start, size_t len);
 
 #endif
