+++ 42-cursus/common-core-rank01/get_next_line/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: rtorrent <rtorrent@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 12:14:46 by rtorrent          #+#    #+#             */
-/*   Updated: 2023/06/30 19:33:22 by rtorrent         ###   ########.fr       */
+/*   Updated: 2023/07/04 17:53:22 by rtorrent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,7 +65,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 	return (src - src0);
 }
 
-char	*ft_substr(const char *s, unsigned int start, size_t len)
+char	*ft_substr(const char *s, size_t start, size_t len)
 {
 	size_t	size;
 	char	*p;
